+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- Fix issue with infinite loop
+- Add abortsignal support
+- Only query first word when string with multiple words is entered
+
 # v1.0.0
 
 - Change result format from just the "result" string returned to be "term,result"